pyproject.toml
[tool.poetry]
name = "openkiwi-tasting"
version = "0.1.0"
description = "OpenKiwi demonstration app."
authors = ["AI Research, Unbabel <openkiwi@unbabel.com>"]
license = "AGPL-3.0"
readme = 'README.md'
homepage = 'https://github.com/Unbabel/OpenKiwi-Tasting'
repository = 'https://github.com/Unbabel/OpenKiwi-Tasting'
keywords = ['OpenKiwi', 'Quality Estimation', 'Machine Translation', 'Unbabel']
packages = [
    {include = "kiwi_tasting"},
]
include = ['pyproject.toml', 'LICENSE']

[tool.poetry.scripts]
kiwi-tasting = 'kiwi_tasting.__main__:run'

[tool.poetry.dependencies]
python = "^3.7"
streamlit = "^0.69.2"
openkiwi = "^2.0.0"
st-annotated-text = "^1.0.1"
pydantic = "^1.5"
omegaconf = "^1.4.1"
transformers = "^3.0.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"
pre-commit = "^2.7.1"

[tool.black]
skip-string-normalization = true  # Don't switch to double quotes
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
